refactor: cleanup @zowe/zos-jobs-for-zowe-sdk

+++ zowe-zos-jobs-for-zowe-sdk/src/jsMain/kotlin/zowe/sdk/zos/jobs/GetJobs.kt
@@ -9,38 +9,26 @@ import zowe.sdk.zos.jobs.doc.input.IGetJobsParms
 import zowe.sdk.zos.jobs.doc.response.IJob
 import zowe.sdk.zos.jobs.doc.response.IJobFile
 
-// unhandled import: AbstractSession from "@zowe/imperative"
-// unhandled import: ICommonJobParms from "./"
-// unhandled import: IGetJobsParms from "./"
-// unhandled import: IJob from "./"
-// unhandled import: IJobFile from "./"
-
 /**
  * Class to handle obtaining of z/OS batch job information
- * @export
- * @class GetJobs
  */
-
 external class GetJobs {
   companion object {
     /**
      * Get jobs (defaults to the user ID of the session as owner)
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @return {Promise<IJob[]>} - promise that resolves to an array of IJob objects (matching jobs)
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @return promise that resolves to an array of IJob objects (matching jobs)
      */
     fun getJobs(session: AbstractSession): Promise<Array<IJob>>
 
     /**
      * Get jobs that match a job name preixl
+     *
      * Defaults to jobs owned by the user ID in the session.
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {string} prefix - job name prefix for which to list jobs. Supports wildcard e.g. JOBNM*
+     * @param session z/OSMF connection info
+     * @param prefix job name prefix for which to list jobs. Supports wildcard e.g. JOBNM*
      *                          returns jobs with names starting with "JOBNM"
-     * @return {Promise<IJob[]>} - promise that resolves to an array of IJob objects (matching jobs)
-     * @memberof GetJobs
+     * @return promise that resolves to an array of IJob objects (matching jobs)
      */
     fun getJobsByPrefix(
       session: AbstractSession,
@@ -49,12 +37,10 @@ external class GetJobs {
 
     /**
      * Get jobs that are owned by a certain user or pattern of users
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {string} owner - owner for which to get jobs. Supports wildcard e.g. IBMU* returns jobs owned by
+     * @param session z/OSMF connection info
+     * @param owner owner for which to get jobs. Supports wildcard e.g. IBMU* returns jobs owned by
      *                         all users whose ID beings with "IBMU"
-     * @return {Promise<IJob[]>} - promise that resolves to an array of IJob objects (matching jobs)
-     * @memberof GetJobs
+     * @return promise that resolves to an array of IJob objects (matching jobs)
      */
     fun getJobsByOwner(
       session: AbstractSession,
@@ -63,14 +49,12 @@ external class GetJobs {
 
     /**
      * Get a list of jobs that match an owner and prefix
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {string} owner - owner for which to get jobs. Supports wildcard e.g. IBMU* returns jobs owned by
+     * @param session z/OSMF connection info
+     * @param owner owner for which to get jobs. Supports wildcard e.g. IBMU* returns jobs owned by
      *                         all users whose ID beings with "IBMU"
-     * @param {string} prefix - prefix for which to get jobs. Supports wildcard e.g. JOBNM*
+     * @param prefix prefix for which to get jobs. Supports wildcard e.g. JOBNM*
      *                          returns jobs with names starting with "JOBNM"
-     * @return {Promise<IJob[]>} - promise that resolves to an array of IJob objects (matching jobs)
-     * @memberof GetJobs
+     * @return promise that resolves to an array of IJob objects (matching jobs)
      */
     fun getJobsByOwnerAndPrefix(
       session: AbstractSession,
@@ -80,11 +64,9 @@ external class GetJobs {
 
     /**
      * Get a list of jobs that match various parameters
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
+     * @param session z/OSMF connection info
      * @param {string}
-     * @return {Promise<IJob[]>} - promise that resolves to an array of IJob objects (matching jobs)
-     * @memberof GetJobs
+     * @return promise that resolves to an array of IJob objects (matching jobs)
      */
     fun getJobsByParameters(
       session: AbstractSession,
@@ -93,11 +75,9 @@ external class GetJobs {
 
     /**
      * Get a single job object from an input job id
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {string} jobid - the job ID for the job for which you want to get status
-     * @return {Promise<IJob>} - promise that resolves to an IJob object from an input jobid
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param jobid the job ID for the job for which you want to get status
+     * @return promise that resolves to an IJob object from an input jobid
      */
     fun getJob(
       session: AbstractSession,
@@ -106,11 +86,9 @@ external class GetJobs {
 
     /**
      * Get jobs filtered by owner and prefix.
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {IGetJobsParms} params - parm object (see IGetJobsParms interface for details)
-     * @return {Promise<IJob[]>} - promise that resolves to an array of IJob objects (matching jobs)
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param params parm object (see IGetJobsParms interface for details)
+     * @return promise that resolves to an array of IJob objects (matching jobs)
      */
     fun getJobsCommon(
       session: AbstractSession,
@@ -119,12 +97,10 @@ external class GetJobs {
 
     /**
      * Get the status and other details (e.g. owner, return code) for a job
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {string} jobname - the job name for the job for which you want to get status
-     * @param {string} jobid - the job ID for the job for which you want to get status
-     * @return {Promise<IJob>} - promise that resolves to an IJob object representing the job
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param jobname the job name for the job for which you want to get status
+     * @param jobid the job ID for the job for which you want to get status
+     * @return promise that resolves to an IJob object representing the job
      */
     fun getStatus(
       session: AbstractSession,
@@ -134,15 +110,14 @@ external class GetJobs {
 
     /**
      * Get the status and other details (e.g. owner, return code) for a job
+     *
      * Alternate version of the API that accepts an IJob object returned by
      * other APIs such as SubmitJobs.  Even though the parameter and return
      * value are of the same type, the IJob object returned will have the
      * current status of the job.
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {IJob} job - job object
-     * @return {Promise<IJob>} - promise that resolves to an IJob object representing the job
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param job job object
+     * @return promise that resolves to an IJob object representing the job
      */
     fun getStatusForJob(
       session: AbstractSession,
@@ -151,11 +126,9 @@ external class GetJobs {
 
     /**
      * Get the status and other details (e.g. owner, return code) for a job
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {ICommonJobParms} parms - parm object (see ICommonJobParms interface for details)
-     * @return {Promise<IJob>} - promise that resolves to an IJob object representing the job
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param parms parm object (see ICommonJobParms interface for details)
+     * @return promise that resolves to an IJob object representing the job
      */
     fun getStatusCommon(
       session: AbstractSession,
@@ -164,12 +137,10 @@ external class GetJobs {
 
     /**
      * Get a list of all spool files for a job.
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {string} jobname - the job name for the job for which you want to get a list of spool files
-     * @param {string} jobid - the job ID for the job for which you want to get a list of spool files
-     * @return {Promise<IJobFile[]>} - promise that resolves to an array of IJobFile objects
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param jobname the job name for the job for which you want to get a list of spool files
+     * @param jobid the job ID for the job for which you want to get a list of spool files
+     * @return promise that resolves to an array of IJobFile objects
      */
     fun getSpoolFiles(
       session: AbstractSession,
@@ -179,13 +150,12 @@ external class GetJobs {
 
     /**
      * Get a list of all job spool files for a job
+     *
      * Alternate version of the API that accepts an IJob object returned by
      * other APIs such as SubmitJobs.
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {IJob} job - the job for which you would like to get a list of job spool files
-     * @return {Promise<IJobFile[]>} - promise that resolves to an array of IJobFile objects
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param job the job for which you would like to get a list of job spool files
+     * @return promise that resolves to an array of IJobFile objects
      */
     fun getSpoolFilesForJob(
       session: AbstractSession,
@@ -194,11 +164,9 @@ external class GetJobs {
 
     /**
      * Get a list of all job spool files for a job.
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {ICommonJobParms} parms - parm object (see for details)
-     * @return {Promise<IJobFile[]>} - promise that resolves to an array of IJobFile objectsl
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param parms parm object (see for details)
+     * @return promise that resolves to an array of IJobFile objectsl
      */
     fun getSpoolFilesCommon(
       session: AbstractSession,
@@ -207,12 +175,10 @@ external class GetJobs {
 
     /**
      * Get JCL from a job.
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {string} jobname - the job name for the job for which you want to retrieve JCL
-     * @param {string} jobid - the job ID for the job for which you want to retrieve JCL
-     * @return {Promise<IJob>} - job document on resolve
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param jobname the job name for the job for which you want to retrieve JCL
+     * @param jobid the job ID for the job for which you want to retrieve JCL
+     * @return job document on resolve
      */
     fun getJcl(
       session: AbstractSession,
@@ -222,13 +188,12 @@ external class GetJobs {
 
     /**
      * Get JCL from a job.
+     *
      * Alternate version of the API that accepts an IJob object returned by
      * other APIs such as SubmitJobs.
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {IJob} job - the job for which you would like to retrieve JCL
-     * @return {Promise<string>} - promise that resolves to JCL content
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param job the job for which you would like to retrieve JCL
+     * @return promise that resolves to JCL content
      */
     fun getJclForJob(
       session: AbstractSession,
@@ -237,11 +202,9 @@ external class GetJobs {
 
     /**
      * Get the JCL that was used to submit a job.
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param {ICommonJobParms} parms - parm object (see ICommonJobParms interface for details)
-     * @return {Promise<string>} - promise that resolves to the JCL content
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param parms parm object (see ICommonJobParms interface for details)
+     * @return promise that resolves to the JCL content
      */
     fun getJclCommon(
       session: AbstractSession,
@@ -250,11 +213,9 @@ external class GetJobs {
 
     /**
      * Get spool content from a job (keeping naming convention patter with this duplication function).
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param jobFile - the spool file for which you want to retrieve the content
-     * @return {Promise<string>} - promise that resolves to the spool content
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param jobFile the spool file for which you want to retrieve the content
+     * @return promise that resolves to the spool content
      */
     fun getSpoolContent(
       session: AbstractSession,
@@ -263,28 +224,24 @@ external class GetJobs {
 
     /**
      * Get spool content from a job using the job name, job ID, and spool ID number from z/OSMF
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param jobname - the job name for the job containing the spool content
-     * @param jobid - the job id for the job containing the spool content
-     * @param spoolId - id number assigned by zosmf that identifies the particular job spool file (DD)
-     * @return {Promise<string>} - promise that resolves to the spool content
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param jobname the job name for the job containing the spool content
+     * @param jobid the job id for the job containing the spool content
+     * @param spoolId id number assigned by zosmf that identifies the particular job spool file (DD)
+     * @return promise that resolves to the spool content
      */
     fun getSpoolContentById(
       session: AbstractSession,
       jobname: String,
       jobid: String,
-      spoolId: Double,
+      spoolId: Int,
     ): Promise<String>
 
     /**
      * Get spool content from a job.
-     * @static
-     * @param {AbstractSession} session - z/OSMF connection info
-     * @param jobFile - the spool file for which you want to retrieve the content
-     * @return {Promise<string>} - promise that resolves to the spool content
-     * @memberof GetJobs
+     * @param session z/OSMF connection info
+     * @param jobFile the spool file for which you want to retrieve the content
+     * @return promise that resolves to the spool content
      */
     fun getSpoolContentCommon(
       session: AbstractSession,
