+++ zowe-zos-jobs-for-zowe-sdk/src/jsMain/kotlin/zowe/sdk/zos/jobs/doc/response/IJob.kt
@@ -74,7 +74,7 @@ external interface IJob {
   /**
    * job phase
    */
-  var phase: Double
+  var phase: Int
 
   /**
    * job phase name
